requirements.txt
django==1.11.6
django-mptt